requirements.txt
names-dataset==1.9.0
nltk==3.4.5
numpy==1.17.2
pyahocorasick==1.4.0
pyenchant==2.0.0
